Create A2 Assignment JVS
Dynamic site visitor data
In your JavaScript code, Create your own object/array structure to store the number of site visitors each month for the last 6 months. Use that object to dynamically create an HTML table to be displayed on the page. You should be able to change the object to see changes on the website.

GDPR
Using JavaScript and DOM manipulation, simulate the ability for a user to "accept cookie usage". Provide a footer banner where a user can click a link to "Accept Cookies." When the user clicks "Accept Cookies", change the content on the banner to display new content in the footer stating: "Cookies were accepted. Would you like to revoke?" and provide a link for the user to revoke. If the user chooses to revoke, use JavaScript to revert to the original banner content. (In bigger applications, you would need to send the user to a page that describes cookie usage and permissions that that the user can can revoke from there.)

Reading assistance
- Add an HTML element that allows a use to see a high-contrast version of the site. This element should act as a "toggle" button. This would entail using high-contrast colours on the site instead of the ones we already used in the CSS assignment. Consider background colours and font colors.
- Add more padding around elements and increase the line height and letter spacing.

Keyboard shortcut
- Add a keyboard shortcut, "Ctrl + A", that would toggle Reading Assistance on and off.

Reminder: All code MUST be your own and not code created by others, found online, or generated by AI. My content should be enough to get you by.